+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/BoxTestGenerated.java
@@ -3050,6 +3050,12 @@ public class BoxTestGenerated extends AbstractBoxTest {
       }
 
       @Test
+      @TestMetadata("SourceMemberInjectionTakesPrecedenceOverDaggerInjector.kt")
+      public void testSourceMemberInjectionTakesPrecedenceOverDaggerInjector() {
+        run("SourceMemberInjectionTakesPrecedenceOverDaggerInjector.kt");
+      }
+
+      @Test
       @TestMetadata("ZeroArgConstructorInjectionGeneratedByAnvil.kt")
       public void testZeroArgConstructorInjectionGeneratedByAnvil() {
         run("ZeroArgConstructorInjectionGeneratedByAnvil.kt");